requirements.txt
extractreq==0.0.9
tabulate
spacy==2.2.0
spacy-streamlit
pywsd==1.1.0
graphviz
XlsxWriter
sklearn
stanfordcorenlp
nltk==3.2.5
lemminflect==0.2.1
openpyxl
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
streamlit-pandas-profiling
tracereq==0.0.3
py-automl